feat: add proper markdown rendering for articles and notes
- Create shared markdown renderer utility with comprehensive comments
- Add markdown rendering to ArticlePage and NoteCard components
- Add CSS styling for markdown content (headers, lists, paragraphs)
- Refactor existing editor components to use shared utility
- Fix header sizing and typography for better readability
- Add drop-cap styling for newspaper aesthetic
- Include security measures (HTML escaping) to prevent XSS
