+++ packages/app/src/components/editor/ArticleWizard.tsx
@@ -11,6 +11,7 @@ import { useCategories } from "@/common/useCategories";
 import { useDustClient } from "@/common/useDustClient";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { stash, tables } from "@/mud/stash";
+import { renderMarkdownToHtml } from "@/utils/markdown";
 
 import Step1 from "./ArticleWizardStep1";
 import Step2 from "./ArticleWizardStep2";
@@ -83,124 +84,7 @@ function deleteDraftLocal(id: string) {
   }
 }
 
-// Minimal markdown renderer used for previews (supports headings, bold, italic, lists, paragraphs)
-function renderMarkdownToHtml(md: string) {
-  const escapeHtml = (s: string) =>
-    s
-      .replace(/&/g, "&amp;")
-      .replace(/</g, "&lt;")
-      .replace(/>/g, "&gt;")
-      .replace(/"/g, "&quot;")
-      .replace(/'/g, "&#39;");
 
-  if (!md) return "";
-  // Normalize line endings
-  const text = md.replace(/\r\n?/g, "\n");
-  const lines = text.split("\n");
-
-  let html = "";
-  let inList = false;
-
-  const flushParagraph = (p: string) => {
-    if (!p) return "";
-    return `<p>${p.replace(/\n/g, "<br />")}</p>`;
-  };
-
-  let paraBuf: string[] = [];
-
-  const pushPara = () => {
-    if (paraBuf.length === 0) return;
-    html += flushParagraph(paraBuf.join("\n"));
-    paraBuf = [];
-  };
-
-  for (let i = 0; i < lines.length; i++) {
-    const ln = lines[i];
-    if (/^\s*([-*])\s+/.test(ln)) {
-      // list item
-      pushPara();
-      if (!inList) {
-        inList = true;
-        html += "<ul>";
-      }
-      const item = ln.replace(/^\s*([-*])\s+/, "");
-      let content = escapeHtml(item);
-      // inline formatting
-      content = content.replace(/\*\*(.+?)\*\*/g, "<strong>$1</strong>");
-      content = content.replace(/\*(.+?)\*/g, "<em>$1</em>");
-      html += `<li>${content}</li>`;
-      continue;
-    } else {
-      if (inList) {
-        html += "</ul>";
-        inList = false;
-      }
-    }
-
-    // Headings
-    const h1 = ln.match(/^\s*#\s+(.*)/);
-    const h2 = ln.match(/^\s*##\s+(.*)/);
-    const h3 = ln.match(/^\s*###\s+(.*)/);
-    if (h1) {
-      pushPara();
-      html += `<h1>${escapeHtml(h1[1])
-        .replace(/\*\*(.+?)\*\*/g, "<strong>$1</strong>")
-        .replace(/\*(.+?)\*/g, "<em>$1</em>")}</h1>`;
-      continue;
-    }
-    if (h2) {
-      pushPara();
-      html += `<h2>${escapeHtml(h2[1])
-        .replace(/\*\*(.+?)\*\*/g, "<strong>$1</strong>")
-        .replace(/\*(.+?)\*/g, "<em>$1</em>")}</h2>`;
-      continue;
-    }
-    if (h3) {
-      pushPara();
-      html += `<h3>${escapeHtml(h3[1])
-        .replace(/\*\*(.+?)\*\*/g, "<strong>$1</strong>")
-        .replace(/\*(.+?)\*/g, "<em>$1</em>")}</h3>`;
-      continue;
-    }
-
-    // Empty line -> paragraph break. Emit a visible spacer div for every
-    // blank line so spacing is noticeable in the preview modal.
-    if (ln.trim() === "") {
-      if (paraBuf.length > 0) pushPara();
-      html += '<div style="height:1rem"></div>';
-      continue;
-    }
-
-    // accumulate into paragraph buffer
-    paraBuf.push(
-      escapeHtml(ln)
-        .replace(/\*\*(.+?)\*\*/g, "<strong>$1</strong>")
-        .replace(/\*(.+?)\*/g, "<em>$1</em>")
-    );
-  }
-
-  if (inList) html += "</ul>";
-  pushPara();
-
-  // Add a drop-cap (first-letter styling) if the rendered HTML starts with a paragraph
-  // and not with a heading. We inject a small <span> with inline styles around the
-  // first visible character of the first paragraph. Also handle the case where the
-  // paragraph starts with a <strong> tag.
-  if (html.startsWith("<p>")) {
-    // Case: <p><strong>X...
-    html = html.replace(
-      /^<p>(\s*)<strong>(\s*)([^<\s])/,
-      '<p>$1<strong>$2<span style="float:left;font-size:3rem;line-height:1;margin-right:0.5rem;">$3</span>'
-    );
-    // Case: <p>X...
-    html = html.replace(
-      /^<p>(\s*)([^<\s])/,
-      '<p>$1<span style="float:left;font-size:3rem;line-height:1;margin-right:0.5rem;">$2</span>'
-    );
-  }
-
-  return html;
-}
 
 export const ArticleWizard: React.FC<Props> = ({
   draftId,
